Create new class for AiModel, rename it and put it for use.
