+++ app/src/main/java/com/hfad/palamarchuksuperapp/data/repository/ChatAiRepositoryImpl.kt
@@ -1,21 +1,21 @@
 package com.hfad.palamarchuksuperapp.data.repository
 
 import android.util.Log
+import com.hfad.palamarchuksuperapp.data.entities.AiModel
 import com.hfad.palamarchuksuperapp.data.entities.MessageAI
 import com.hfad.palamarchuksuperapp.data.services.GeminiApiHandler
 import com.hfad.palamarchuksuperapp.data.services.GroqApiHandler
 import com.hfad.palamarchuksuperapp.data.services.OpenAIApiHandler
 import com.hfad.palamarchuksuperapp.domain.models.AppError
 import com.hfad.palamarchuksuperapp.domain.repository.ChatAiRepository
-import com.squareup.moshi.Json
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.update
-import kotlinx.serialization.Serializable
 import javax.inject.Inject
 import com.hfad.palamarchuksuperapp.domain.models.Result
+import com.hfad.palamarchuksuperapp.domain.repository.AiModelHandler
 
 
 class ChatAiRepositoryImpl @Inject constructor(
@@ -62,7 +62,7 @@ class ChatAiRepositoryImpl @Inject constructor(
         }
     }
 
-    override suspend fun getModels(): List<AiModels> {
+    override suspend fun getModels(): List<AiModel> {
 
         when (val response = geminiApiHandler.getModels()) {
             is Result.Success -> {
@@ -87,48 +87,5 @@ class ChatAiRepositoryImpl @Inject constructor(
     }
 
     private val currentModel: AiModels = AiModels.GeminiModels.BASE_MODEL
-
-}
-
-interface AiModels {
-
-    val modelName: String
-
-    @Serializable
-    data class GroqModel(
-        @Json(name = "name")
-        override val modelName: String = "llama-3.2-11b-vision-preview",
-    ) : AiModels
-
-    @Serializable
-    data class GeminiModel(
-        override val modelName: String = "gemini-1.5-flash",
-        val version: String = "1.0.0",
-        val displayName: String = "Gemini",
-        val description: String = "Gemini is a language model that can generate images using the LLM",
-        val supportedGenerationMethods: List<String> = emptyList(),
-        val isSupported: Boolean = supportedGenerationMethods.contains("generateContent"),
-    ) : AiModels
-
-    @Serializable
-    data class OpenAIModel(
-        override val modelName: String = "openai-1",
-    ) : AiModels
-
-
-    enum class GroqModels(override val modelName: String) : AiModels {
-        BASE_MODEL("llama-3.2-11b-vision-preview"),
-        TEXT_MODEL("llama3-groq-8b-8192-tool-use-preview")
-    }
-
-    enum class GeminiModels(override val modelName: String) : AiModels {
-        BASE_MODEL("gemini-1.5-flash"),
-        GEMINI_IMAGE("")
-    }
-
-    enum class OpenAIModels(override val modelName: String) : AiModels {
-        BASE_MODEL(""),
-        OPENAI_IMAGE("")
-    }
 
 }
